switch-snmp/requirements.txt
python-dotenv
influxdb-client
pandas